Move openDatabricksConfigFile to utils
Call it from relevant places directly, since the command might not be registered
